XW-2103 | move search page object to elements directory
