+++ src/test/java/com/wikia/webdriver/elements/oasis/pages/SearchPage.java
@@ -1,4 +1,4 @@
-package com.wikia.webdriver.pageobjectsfactory.pageobject.search;
+package com.wikia.webdriver.elements.oasis.pages;
 
 import com.wikia.webdriver.common.logging.PageObjectLogging;
 import com.wikia.webdriver.elements.common.Navigate;
@@ -7,7 +7,7 @@ import com.wikia.webdriver.pageobjectsfactory.pageobject.WikiBasePageObject;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
-public class SearchPagePageObject extends WikiBasePageObject {
+public class SearchPage extends WikiBasePageObject {
 
   @FindBy(css = "#search-v2-input")
   private WebElement searchInput;
@@ -21,24 +21,24 @@ public class SearchPagePageObject extends WikiBasePageObject {
   private static final String SPECIAL_SEARCH_PAGE = "/wiki/Special:Search";
   private static final String EXISTING_COMMUNITY_NAME = "pokemon";
 
-  public SearchPagePageObject() {
+  public SearchPage() {
     super();
   }
 
-  public SearchPagePageObject navigateToSearchPage() {
+  public SearchPage navigateToSearchPage() {
     new Navigate().toPage(SPECIAL_SEARCH_PAGE);
 
     return this;
   }
 
-  public SearchPagePageObject typeInSearchInputCommunityName() {
+  public SearchPage typeInSearchInputCommunityName() {
     wait.forElementClickable(searchInput);
     searchInput.sendKeys(EXISTING_COMMUNITY_NAME);
 
     return this;
   }
 
-  public SearchPagePageObject clickSearchButton() {
+  public SearchPage clickSearchButton() {
     wait.forElementClickable(searchButton);
     searchButton.click();
